move project delete to workspace
