+++ lib/api/v3/projects/projects_api.rb
@@ -65,9 +65,6 @@ module API
             end
 
             get &::API::V3::Utilities::Endpoints::Show.new(model: Project).mount
-            delete &::API::V3::Utilities::Endpoints::Delete.new(model: Project,
-                                                                process_service: ::Projects::ScheduleDeletionService)
-                                                           .mount
             mount API::V3::Projects::Copy::CopyAPI
 
             mount ::API::V3::Workspaces::NestedApis
